pyproject.toml
[tool.poetry]
name = "shop-app"
version = "0.1.0"
description = ""
authors = ["AlexSokolov <speaker04@mail.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
diploma-frontend = {path = "diploma-frontend/dist/diploma-frontend-0.6.tar.gz"}
django = "^5.1.4"
djangorestframework = "^3.15.2"
pillow = "^11.0.0"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"
flake8 = "^7.1.1"
flake8-bugbear = "^24.8.19"
flake8-pie = "^0.16.0"
django-stubs = {extras = ["compatible-mypy"], version = "^5.1.1"}
psycopg2-binary = "^2.9.10"
djangorestframework-stubs = "^3.15.1"
drf-spectacular = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main"
]
disable_error_code = ["attr-defined",]
#ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "shop_app"
